Remove '.class' for 'through_options'
A student was confused by this. They correctly pointed out that since we're going to use the Associatable module by extending a class, self.class would be Class. In the solutions, this line is moved into define_method, but it also works to define through_options before that if we remove the '.class.'